Revert "Refactor suggestion logic into module"
